Configure storage node in DataMetadataEndpoints.test
